+++ plugins/notebook/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - add an opt-in local Microsandbox microVM execution backend while retaining Docker as the default
+- add an opt-in Daytona notebook backend pinned to the deployed legacy SDK contract
 
 ## [0.2.0](https://github.com/SarthakJariwala/sqlsaber/compare/sqlsaber-notebook-v0.1.0...sqlsaber-notebook-v0.2.0) (2026-07-22)